refactor: move spectrum view into ui module
