+++ rigflow-client/src/ui/spectrum_view.rs
